move kafka event handlers from service folder to events folder
