+++ src/services/index.js
@@ -2,8 +2,8 @@
 
 import config from 'config';
 import RabbitMQService from './rabbitmq';
-import startKafkaConsumer from './kafka/kafkaConsumer';
-import kafkaHandler from './kafka/kafkaHandlers';
+import startKafkaConsumer from './kafkaConsumer';
+import { kafkaHandlers } from '../events';
 
 /**
  * Responsible for establishing connections to all external services
@@ -33,7 +33,7 @@ module.exports = (fapp, logger) => {
     .then(() => {
       logger.info('RabbitMQ service initialized');
     })
-    .then(() => startKafkaConsumer(kafkaHandler, app, logger))
+    .then(() => startKafkaConsumer(kafkaHandlers, app, logger))
     .then(() => {
       logger.info('Kafka consumer service initialized');
     })
