+++ src/services/rabbitmq.js
@@ -1,7 +1,7 @@
 /* globals Promise */
 import _ from 'lodash';
 import amqplib from 'amqplib';
-import handlers from '../events';
+import { rabbitHandlers as handlers } from '../events';
 
 module.exports = class RabbitMQService {
